Okay, have rewritten a lot of the error recovery code.  Added some code in the Plan to identify dead-ends, but, a few hours later, I finally realized what I've been missing for two months: the reason the position signatures were not doing the job is that they were *too* specific.  The only thing that matters is the State number, and the source extent (@node start_position to follow_position).  You should never re-enter the same state with the same source extent, because, if you do, you will begin repeating EVERYTHING!  This is, in other words, the solution to all sorts of problems, including eliminating trying both "x + y" and "x - y" as recovery options for "x y".  Doh!
So, this commit is to capture all the code I'm about to discard.  The code mostly works, right now, but still ends up in infinite loops more often than it should.  There are also a bunch of subtle bugs with error recovery stuff.  In other words, it runs, but it doesn't work.  Heave to.


git-svn-id: http://courage.codebasehq.com/rcc/archive.svn/trunk@35 6f01870a-c388-4d8b-a67f-fe219a3627e8